Extract semaphore operations to their own file
